refactor: (#281) domain 의존으로 인한 Web Enum 제거
