+++ simtong-presentation/src/main/kotlin/team/comit/simtong/domain/holiday/dto/response/QueryIndividualHolidaysWebResponse.kt
@@ -1,16 +1,17 @@
-package team.comit.simtong.domain.holiday.dto
+package team.comit.simtong.domain.holiday.dto.response
 
+import team.comit.simtong.domain.holiday.model.value.HolidayType
 import java.time.LocalDate
 
 /**
  *
- * 개인 휴무일 정보를 전송하는 QueryIndividualHolidayResponse
+ * 개인 휴무일 정보를 전송하는 QueryIndividualHolidayWebResponse
  *
  * @author Chokyunghyeon
  * @date 2022/12/05
- * @version 1.0.0
+ * @version 1.2.5
  **/
-data class QueryIndividualHolidaysResponse(
+data class QueryIndividualHolidaysWebResponse(
     val holidays: List<IndividualHolidayResponse>
 )
 
@@ -24,5 +25,5 @@ data class QueryIndividualHolidaysResponse(
  **/
 data class IndividualHolidayResponse(
     val date: LocalDate,
-    val type: String
+    val type: HolidayType
 )
